requirements.txt
antlr4-python3-runtime==4.9.3
bounded-pool-executor==0.0.3
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
cssselect2==0.6.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
defusedxml==0.7.1
epam.indigo==1.7.4
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1661381345715/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1661444073798/work
hydra-core==1.2.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1657953096987/work
lxml==4.9.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1661439866112/work
munkres==1.1.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1660673040956/work
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1660385859913/work
pqdm==0.2.0
pycairo==1.21.0
pycparser==2.21
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1660580468234/work
PyYAML==6.0
reportlab==3.5.68
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1659998781104/work
svglib==1.4.1
termcolor==1.1.0
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.3.0
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1649111910890/work
webencodings==0.5.1
wget==3.2

matplotlib~=3.5.3
numpy~=1.23.2
Pillow~=9.2.0